% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract2.R
\name{array_type}
\alias{array_type}
\title{Simplified array data type of R to match Stan's parameter dimensions
Vector: 1D;  matrix: 2D;  array: 3D or above}
\usage{
array_type(x)
}
\description{
Simplified array data type of R to match Stan's parameter dimensions
Vector: 1D;  matrix: 2D;  array: 3D or above
}
