% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{as_c_num}
\alias{as_c_num}
\alias{as_c_chr}
\title{Vector object as R input string}
\usage{
as_c_num(x)

as_c_chr(x)
}
\description{
Vector object as R input string
}
\examples{
as_c_num(1:3)
as_c_chr(1:3)
}
