% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{as_vec}
\alias{as_vec}
\title{Handy vector construction from string}
\usage{
as_vec(x, cast_num = TRUE)
}
\arguments{
\item{x}{String. A string to construct a vector from (see examples).
If \code{x} has length > 1, \code{stom::as_c_num} or \code{stom::as_c_chr}
would be called (depending on input type) to print out the
concatenation syntax \code{c(...)} in the console.}

\item{cast_num}{Boolean. Whether to convert string to numeric when
applicable. See examples.}
}
\description{
Handy vector construction from string
}
\examples{
# Construct vector from string
as_vec("a b c")
as_vec("d, e  f")   # Commas have precedence over spaces

# Auto typecasting
as_vec("1 2 -3.1")
as_vec("04, 05")
as_vec("04, 05", cast_num=FALSE)  # Disable typecasting
}
