% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{cmd_get_value}
\alias{cmd_get_value}
\alias{cmd_has}
\alias{cmd_parse_date}
\title{Primitive parsing of command line arguments}
\usage{
cmd_get_value(args, pre)

cmd_has(args, has)

cmd_parse_date(args, pre = "--date=")
}
\arguments{
\item{args}{Character vector. Returned by \code{base::commandArgs(TRUE)}.}

\item{pre}{String. Command line prefix to retrieve value from.}

\item{has}{Character vector. Arguments to be checked for presence in the
command line arguments.}
}
\description{
Primitive parsing of command line arguments
}
\examples{
args = c( "--date=2022-11-20", "--second" )
cmd_get_value(args, "--date=")
cmd_parse_date(args)
cmd_has(args, "--second")
}
