% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{cmd_parse_args}
\alias{cmd_parse_args}
\alias{cmd_parse_key_val}
\title{(Deprecated) Full parsing of command line arguments}
\usage{
cmd_parse_args(x)

cmd_parse_key_val(x)
}
\arguments{
\item{x}{A vector of command line arguments.}
}
\value{
A list of parsed arguments
}
\description{
(Deprecated) Full parsing of command line arguments
}
\examples{
x = c("--date=2021", "--hi-w=y", "--hi_w=y", "./file-info.txt", "--hi")
cmd_parse_args(x)
}
