% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{col.alpha}
\alias{col.alpha}
\alias{shade}
\title{Functions stolen from the rethinking package}
\usage{
col.alpha(acol, alpha = 0.5)

shade(
  object,
  lim,
  label = NULL,
  col = col.alpha("black", 0.15),
  border = NA,
  ...
)
}
\description{
Functions stolen from the rethinking package
}
