% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{curve_mvnorm}
\alias{curve_mvnorm}
\title{Add multivariate normal contour to plot}
\usage{
curve_mvnorm(
  mu,
  Sigma,
  level = c(seq(0.1, 0.9, 0.1), 0.99),
  col = 1,
  alpha = 0.1,
  ...
)
}
\arguments{
\item{mu}{Vector of MVNormal center.}

\item{Sigma}{Covariance matrix of MVNormal.}

\item{level}{The values the of cumulative density to plot contours.}

\item{col}{Color of the contours.}

\item{alpha}{Opacity of the contours.}

\item{...}{Further arguments passed to \code{\link[graphics:lines]{graphics::lines()}}.}
}
\description{
Add multivariate normal contour to plot
}
\examples{
mu = c(.5, .7)
Rho = matrix(c(
    1, .2,
   .2,  1 ), nrow=2, byrow=T)
s = c( 1.5, 1.2 )
Sigma = diag(s) \%*\% Rho \%*\% diag(s)
x = MASS::mvrnorm(50, mu, Sigma)
plot(x)
curve_mvnorm(mu, Sigma)
}
