% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{erf}
\alias{erf}
\alias{inv_erf}
\alias{tanh}
\alias{inv_tanh}
\title{Links for mapping reals to values between -1 and 1.}
\usage{
erf(x)

inv_erf(x)

tanh(x)

inv_tanh(x)
}
\arguments{
\item{x}{Double.}
}
\description{
Links for mapping reals to values between -1 and 1.
}
\details{
\code{erf()}: \url{https://en.wikipedia.org/wiki/Error_function}.
}
