% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{export_docs}
\alias{export_docs}
\alias{export_docs_pdf}
\title{Extract docstring from R script}
\usage{
export_docs(path, outfp = NULL)

export_docs_pdf(path, outfp)
}
\arguments{
\item{path}{Path to input R script.}

\item{outfp}{Output file path to write to.}
}
\description{
Extract docstring from R script
}
