% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.R
\name{fstem}
\alias{fstem}
\title{R equivalent of Python pathlib's \code{Path.stem}}
\usage{
fstem(x)
}
\arguments{
\item{x}{Character. File name.}
}
\description{
R equivalent of Python pathlib's \code{Path.stem}
}
