% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract2.R
\name{get_sim_param_info}
\alias{get_sim_param_info}
\title{Get parameter information of a simulation}
\usage{
get_sim_param_info(lst)
}
\arguments{
\item{lst}{List. A list of parameter values from the simulation. Refer
to \code{system.file("cases", "treatment_dynamics", "2-varying-effect-model", "simulation.R", package="stom")}
for the data structure of the input. It should match the structure of
\code{sim_data()$params}.}
}
\value{
A list of parameters with their dimension info.
}
\description{
The input list can be nested in arbitrary levels, as long as the names at the
last level matches those used in the corresponding Stan model.
}
\examples{
fp = system.file("cases", "treatment_dynamics", "1-simplified-normal-model", "simulation.R", package="stom")
source(fp)
d = sim_data()
str(d$params)
get_sim_param_info(d$params)
}
