% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan2r.R
\name{get_stan_param_info}
\alias{get_stan_param_info}
\title{Infer parameter data structure from Stan's parameter names}
\usage{
get_stan_param_info(params)
}
\arguments{
\item{params}{Character. A vector of parameter names to pass in.}
}
\value{
A list of parameters, matching their input order, with data structure
information.
}
\description{
This function is used in \code{stom::extract2()} for inferring parameter structure
from Stan model outputs.
}
\examples{
p_scalar = c( "alpha", "beta" )
p_1d = c( "x[1]", "x[2]" )
p_2d = c("y[1,1]", "y[2,1]", "y[1,2]", "y[2,2]")
p_3d = c("z[1,1,1]", "z[2,1,1]", "z[1,2,1]", "z[2,2,1]", "z[1,1,2]", "z[2,1,2]", "z[1,2,2]", "z[2,2,2]")
params = c( p_scalar, p_1d, p_2d, p_3d )
get_stan_param_info(params)
}
