% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{ignore_dollars_with_begin_equation}
\alias{ignore_dollars_with_begin_equation}
\title{Spcial function to deal with conflicts between \verb{$$} and equation env in latex}
\usage{
ignore_dollars_with_begin_equation(lines)
}
\description{
Spcial function to deal with conflicts between \verb{$$} and equation env in latex
}
