% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick.R
\name{image_combine}
\alias{image_combine}
\title{Joining multiple SVG/PNG images into a single one}
\usage{
image_combine(
  fps,
  output = NULL,
  stack = FALSE,
  height = NULL,
  width = NULL,
  sep_fct = 0.1,
  background = "transparent"
)
}
\arguments{
\item{fps}{A character vector of paths to the images to be combined.}

\item{output}{The output file path. If \code{NULL}, the image is printed.}

\item{stack}{Logical. \code{TRUE} joins images in a top-to-bottom direction.
\code{FALSE} joins images in a left-to-right direction.}

\item{width, height}{An integer specifying the common widths and/or heights of
the images to be read in. Works only with SVG images. For joining the
images side-by-side, setting a common height is recommended, whereas
when the images are stacked, setting a common width results in a more
desirable output image.}

\item{sep_fct}{Numeric. \code{sep_fct} determines the space between the joined
images. The space between the joined images is determined by:
\deqn{\textrm{Horizontal space} = \textrm{sep_fct} \times \textrm{max}(\textrm{widths of input images}) }
\deqn{\textrm{Vertical space}   = \textrm{sep_fct} \times \textrm{max}(\textrm{heights of input images}) }}

\item{background}{Character. The color to add to background. By default, "transparent".}
}
\description{
Joining multiple SVG/PNG images into a single one
}
