% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_format_conversion.R
\name{inkscape_pdf_pager}
\alias{inkscape_pdf_pager}
\title{Construct output file paths based on input PDF page numbers}
\usage{
inkscape_pdf_pager(pdf_file, to_ext = ".svg", outfile = NULL, page = NULL)
}
\description{
Construct output file paths based on input PDF page numbers
}
