% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inkscape.R
\name{inkscape_run_through_pages}
\alias{inkscape_run_through_pages}
\title{Run Inkscape through pages of PDF input, where page info is given by the
\code{out_info} arg obtained through \code{inkscape_pdf_pager()}}
\usage{
inkscape_run_through_pages(infile, base_cmd, out_info)
}
\description{
Run Inkscape through pages of PDF input, where page info is given by the
\code{out_info} arg obtained through \code{inkscape_pdf_pager()}
}
