% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{inline}
\alias{inline}
\alias{inline_zh}
\title{Format vectors as inline text}
\usage{
inline(x, sep = ", ", last_sep = ", and ", two_sep = " and ")

inline_zh(x, sep = "<U+3001>", last_sep = "<U+3001>", two_sep = " <U+8207> ")
}
\description{
These functions are expect to work with R Markdown's inline R chunk syntax
\code{`r `}, such that, for instance,
\verb{they are numbered `r inline(1:3)` respectively} results in
\verb{they are numbered 1, 2, and 3 respectively}.
}
\examples{
inline(1:2)
inline(1:3)
inline_zh(1:2)
inline_zh(1:3)
}
