% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{log_mix}
\alias{log_mix}
\title{log_mix from stan}
\usage{
log_mix(theta, lp1, lp2, log_lp = TRUE)
}
\arguments{
\item{theta, lp1, lp2}{Numeric. See
\url{https://mc-stan.org/docs/2_21/functions-reference/composed-functions.html}
for the documentation.}

\item{log_lp}{Logical. Defaults to TRUE, which complies with Stan's
\code{log_mix()}. If \code{FALSE}, \code{lp1} and \code{lp2} is further converted to
logarithms before the mixture. In such cases, the inputs
\code{theta}, \code{lp1}, and \code{lp2} should all be in the probability space
to make sense.}
}
\description{
log_mix from stan
}
\examples{
theta = .3
lp1 = .99
lp2 = .1
log_mix(theta, lp1, lp2, log_lp=FALSE)
}
