% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistics.R
\name{logit}
\alias{logit}
\alias{inv_logit}
\alias{logistic}
\title{Links and Inverse Links for Logistic Regression}
\usage{
logit(p)

inv_logit(x)

logistic(x)
}
\arguments{
\item{p}{Numeric vector. Values on the probability scale.}

\item{x}{Numeric vector. Values on the linear scale.}
}
\description{
Links and Inverse Links for Logistic Regression
}
