% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{lst2message}
\alias{lst2message}
\title{Print named lists as formatted messages}
\usage{
lst2message(lst, indent = 0L, message = T, item_sep = "\\n")
}
\arguments{
\item{lst}{A list. Input to be printed.}

\item{indent}{Integer. The number of spaces padded on the left
of the output.}

\item{message}{Logical. Whether to print the input as message.
Defaults to \code{TRUE}. If \code{FALSE}, returns a character vector
with formatted content.}

\item{item_sep}{A string. The separator between the list elements.
Defaults to \verb{\\n}.}

\item{name}{description}
}
\description{
Print named lists as formatted messages
}
\examples{
lst2message(list(xyz = 1, xy = 2:3))
message("[Message]")
#' lst2message(list(xyz = 1, xy = 2:3), indent=3)
lst2message(list(xyz = 1, xy = 2:3), message=F)
}
