% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_format_conversion.R
\name{mutool_convert}
\alias{mutool_convert}
\title{Convert PDF to images by mutool}
\usage{
mutool_convert(
  infile,
  outfile = NULL,
  format = c("png", "svg"),
  page = NULL,
  dpi = 300,
  white = FALSE
)
}
\arguments{
\item{infile}{String. Path to input PDF file.}

\item{outfile}{String Output file path. Defaults to NULL, and is inferred from \code{infile}.}

\item{format}{String. Output format. Ignored when \code{outfile} is given.}

\item{page}{Integer vector. The pages (starting with 1) to extract from the
PDF file, defaults to \code{NULL}, in which every page is converted to
an image.}

\item{dpi}{Integer. "Dots Per Inch" for PNG outputs.}
}
\description{
Convert PDF to images by mutool
}
