% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{n_samples}
\alias{n_samples}
\title{Get number of samples (chains combined) from CmdStanFit}
\usage{
n_samples(fit)
}
\arguments{
\item{fit}{CmdStanFit object.}
}
\value{
Integer
}
\description{
Get number of samples (chains combined) from CmdStanFit
}
