% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{pandoc_pdf}
\alias{pandoc_pdf}
\alias{pandoc_html}
\title{Pandoc wrappers for converting markdown to pre-styled documents}
\usage{
pandoc_pdf(fin, outfp, style = "amsart", ...)

pandoc_html(fin, outfp, args = NULL, style = NULL)
}
\description{
Pandoc wrappers for converting markdown to pre-styled documents
}
