% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{pat_depth}
\alias{pat_depth}
\title{Stan parameter regex pattern}
\usage{
pat_depth(depth = 1)
}
\arguments{
\item{depth}{Integer. The parameters' dimension number. Parameters with
dimension number equal to or less than \code{depth} will be retrieved.}
}
\description{
Stan parameter regex pattern
}
\examples{
pat_depth(0)
pat_depth(1)
pat_depth(2)
grepl( pat_depth(0), c("a", "a[1]", "a[1,2]", "a[1,2,3]") )
grepl( pat_depth(2), c("a", "a[1]", "a[1,2]", "a[1,2,3]") )
}
