% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseR-forests.R
\name{plot_forest}
\alias{plot_forest}
\title{Forest plot for multiple densities}
\usage{
plot_forest(
  dat,
  labels,
  shade_inv = c(0.25, 0.75),
  center = c("mean", "median", "mode"),
  vert_ref = NULL,
  col = 2,
  xlim = NULL,
  sep_fct = 0.5,
  adj = -0,
  ...
)
}
\arguments{
\item{dat}{A list of numeric vectors with each representing
a set of observations from a distribution.}

\item{labels}{Character vector. Labels plotted on the y-axis.}

\item{shade_inv}{Numeric vector of length 2. The central area to
be shaded in the distributions By default, the central 50\%
regions are shaded.}

\item{center}{String. The statistic used for drawing the central
tendency line. Defaults to \code{mean}.}

\item{vert_ref}{Numeric. The position of the vertical reference line.
If NULL, nothing is plotted.}

\item{col}{Integer or String. A color code recognized by base R.
Determines the color of the central tendency lines and
shaded regions.}

\item{xlim}{Numeric. A vector of length 2 giving the plotted x range.}

\item{sep_fct, adj}{Parameters to tweak the positions of the distributions.
\code{sep_fct} changes the distances between a pair of distributions.
\code{adj} shifts the baselines of the distributions, with positive
values shifting them upwards and negative values downwards.}

\item{...}{Additional arguments passed to the main \code{plot()} function.}
}
\description{
Forest plot for multiple densities
}
\examples{
# Set up the data
set.seed(100)
data1 <- rnorm(1e6, 0, 1.5)
data2 <- rnorm(1e6, 1, 1.5)
data3 <- rnorm(1e6, -1, 1.2)
data4 <- rnorm(1e6, 1.5, 1.8)

plot_forest(list(data1,data2,data3,data4), labels=LETTERS[1:4],
            vert_ref = 0, center="mode", xlim=c(-10,10),
            main = "A comparision of four Normal distributions")

}
