% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom-data.R
\name{prepare_drive_model}
\alias{prepare_drive_model}
\title{Gather Stan models for Google Drive upload}
\usage{
prepare_drive_model()
}
\description{
Gather Stan models for Google Drive upload
}
\seealso{
\code{\link[=save_model]{save_model()}}
}
