% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-detection.R
\name{prob_input}
\alias{prob_input}
\alias{confusion_input}
\title{Sensitivity and Specificity calculation}
\usage{
prob_input(sensitivity, specificity, prevalence)

confusion_input(TP, TN, FP, FN)
}
\arguments{
\item{sensitivity}{Numeric. P(+|P), the probability of a positive test result
given that the result is truly positive.}

\item{specificity}{Numeric. P(-|N), the probability of a negative test result
given that the result is truly negative.}

\item{prevalence}{Numeric. P(P), the probability of truly positive results in
the population.}

\item{TP}{Integer. True positive count.}

\item{TN}{Integer. True negative count.}

\item{FP}{Integer. False positive count.}

\item{FN}{Integer. False negative count.}
}
\description{
Sensitivity and Specificity calculation
}
\examples{
post = prob_input(sensitivity=.9, specificity=.97, prevalence=.008)
post

# Assume there are 2957 positive cases and 77043 negative cases.
# The code below computes TP, FP, TN, FN
counts = post * c(2957, 2957, 77043, 77043)
counts

confusion_input(TP=576, FP=2381, TN=76979, FN=64)
}
