% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rbern}
\alias{rbern}
\title{Bernoulli Random Number Generator}
\usage{
rbern(prob, n = length(prob))
}
\arguments{
\item{prob}{Vector of probabilities.}

\item{n}{Numbers of sample, defaults to the length of \code{prob}.}
}
\description{
Alias of \code{rbinom()} for a single trial (\code{size = 1}).
}
