% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{replace_file_ext}
\alias{replace_file_ext}
\title{Replace file extension}
\usage{
replace_file_ext(fp, ext = "")
}
\arguments{
\item{fp}{Character vector of file paths.}

\item{ext}{String. Extension (with dot) to change to.}
}
\description{
Replace file extension
}
\examples{
replace_file_ext("x/y/zcda.txt")
replace_file_ext(paste0(1:3, ".txt"), ".stan")
}
