% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rgampois2}
\alias{rgampois2}
\title{Random Number Generator from Gamma-Poisson Distribution}
\usage{
rgampois2(n, mu, phi)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Poisson mean (rate)}

\item{phi}{Parameter determining the dispersion of the outcomes.
This parameter is equivalent to Stan's \code{neg_binomial_2(mu, phi)}'s \code{phi}.}
}
\description{
Random Number Generator from Gamma-Poisson Distribution
}
\details{
See Chapter 12.1.2 of Statistical Rethinking 2th Edition, which details the
Gamma-Poisson distribution for modeling over-dispersed count data.
}
