% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rlkjcorr}
\alias{rlkjcorr}
\title{Random Number Generator for Cholesky LKJ Correlation Distribution}
\usage{
rlkjcorr(n, K, eta = 1)
}
\arguments{
\item{n}{Number of samples.}

\item{K}{Dimension of matrix}

\item{eta}{Parameter controling the strengths of correlations}
}
\description{
Random Number Generator for Cholesky LKJ Correlation Distribution
}
\details{
Function taken from https://github.com/rmcelreath/rethinking/blob/master/R/distributions.r#L214
}
