% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rordlogit}
\alias{rordlogit}
\title{Random Number Generator for ordered categories}
\usage{
rordlogit(phi, cutpoints)
}
\arguments{
\item{phi}{Numeric. The linear term(s) to subtract from
each ordered categories.}

\item{cutpoints}{A vector of cutpoints for delineating the
ordered categories on the logit scale. For generating
\code{n} categories, \code{n-1} cutpoints are needed. Do not
include (negative) \code{Inf} values.}
}
\description{
Random Number Generator for ordered categories
}
\examples{
cutpoints = c( -1.5, 0, 1.5 )
rordlogit( phi=0, cutpoints=cutpoints )
x = replicate( 1e5, {rordlogit(0, cutpoints)})
table(x) / length(x)
logistic( c(cutpoints, Inf) ) - logistic( c(-Inf, cutpoints) )
}
