% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rtcauchy}
\alias{rtcauchy}
\title{Random Number Generator for Truncated Cauchy Distribution}
\usage{
rtcauchy(n, location = 0, scale = 1, lower = 0, upper = Inf)
}
\arguments{
\item{n}{Number of samples.}

\item{location, scale}{Parameters of the Cauchy distribution, see \code{stats::rcauchy}.}

\item{lower}{The lower bound for the truncation.}

\item{upper}{The upper bound for the truncation.}
}
\description{
The parameters are set to generate numbers from a half-normal distribution
by default.
}
