% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rtnorm}
\alias{rtnorm}
\title{Random Number Generator for Truncated Normal Distribution}
\usage{
rtnorm(n, m = 0, s = 1, lower = 0, upper = Inf)
}
\arguments{
\item{n}{Number of samples.}

\item{m}{The mean of the untruncated normal distribution.}

\item{s}{The standard deviation of the untruncated normal distribution.}

\item{lower}{The lower bound for truncation.}

\item{upper}{The upper bound for truncation.}
}
\description{
The parameters are set to generate numbers from a half-normal distribution
by default.
}
