% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{save_model}
\alias{save_model}
\title{Save \code{CmdStanFit} object as RDS with a cross-platform consistent name}
\usage{
save_model(m, fp = NULL)
}
\arguments{
\item{m}{A \code{CmdStanFit} object returned from \code{stom::stan()}.}
}
\description{
Save \code{CmdStanFit} object as RDS with a cross-platform consistent name
}
\seealso{
\code{\link[=prepare_drive_model]{prepare_drive_model()}}
}
