% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{softmax}
\alias{softmax}
\title{Numerically stable softmax function}
\usage{
softmax(x)
}
\arguments{
\item{x}{Numeric. A vector of reals to map to probabilities.}
}
\description{
Numerically stable softmax function
}
\examples{
softmax(1:3)
softmax(1:3 + 9999999)
softmax(c(12345, 67890, 99999999))
}
