% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{stan}
\alias{stan}
\title{Compile and sample with cmdstanr}
\usage{
stan(stan_file, data, ...)
}
\arguments{
\item{stan_file}{String. Stan file path.}

\item{data}{List. Data list.}

\item{seed}{Integer. Random seed passed to cmdstanr's \verb{$sample()}.}
}
\value{
A \code{CmdStanFit} object.
}
\description{
Compile and sample with cmdstanr
}
