% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{stan2bin}
\alias{stan2bin}
\title{Convert stan file path to stan binary file path}
\usage{
stan2bin(fp)
}
\arguments{
\item{fp}{Path to \verb{*.stan} file}
}
\description{
Assumes \verb{*.stan} & compiled binary found in same directory.
}
