% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector to fixed mean and std}
\usage{
standardize(x, m = 0, s = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{m}{Mean.}

\item{s}{Standard deviation.}
}
\description{
Standardize a vector to fixed mean and std
}
