% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom-package.R
\docType{package}
\name{stom-package}
\alias{stom}
\alias{stom-package}
\title{stom: Miscellaneous Functions for Statistical Modeling}
\description{
Personal package for a collection of functions commonly used for statistical modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://yongfu.name/stom}
  \item \url{https://github.com/liao961120/stom}
}

}
\author{
\strong{Maintainer}: Yongfu Liao \email{liao961120@gmail.com} (\href{https://orcid.org/0000-0002-1814-2993}{ORCID})

}
\keyword{internal}
