% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{test_cmdstanr}
\alias{test_cmdstanr}
\title{Test cmdstanr compilation}
\usage{
test_cmdstanr(clean = F, draw_sample = T)
}
\arguments{
\item{clean}{Whether to recompile if binary found, defaults to FALSE.}

\item{draw_sample}{Whether to draw samples, defaults to TRUE.}
}
\description{
Test cmdstanr compilation
}
