% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexical_diversity.R
\name{TTR}
\alias{TTR}
\alias{MATTR}
\alias{MSTTR}
\alias{HDD}
\alias{ATTR}
\alias{MTLD}
\alias{YuleK}
\title{Lexical diversity measures}
\usage{
TTR(x)

MATTR(x, window = 50)

MSTTR(x, seg.len = 50)

HDD(x, drawn = 42)

ATTR(x, drawn = 42)

MTLD(x, factor.size = 0.72)

YuleK(x)
}
\description{
Lexical diversity measures
}
