% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vals2cols}
\alias{vals2cols}
\title{Convert numerical values to divergent colors}
\usage{
vals2cols(x, center = 0)
}
\arguments{
\item{x}{Numeric.}

\item{center}{Integer. Currently only \code{0} is valid.
If \code{center == 0}, \code{0} is used as the center and corresponds to
the central color (lightgoldenrod1).
Otherwise, \code{x} must have values in the range \link{0, 1}.}
}
\value{
A vector of hex colors.
}
\description{
Convert numerical values to divergent colors
}
\examples{
x = rnorm(1e3)
plot(seq(x), x, col= vals2cols(x), pch=19)
}
