% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{year}
\alias{year}
\alias{month}
\alias{day}
\title{Extract Year/Month/Day from date string}
\usage{
year(x)

month(x)

day(x)
}
\arguments{
\item{x}{Character or Date vector.}
}
\value{
Numeric vector
}
\description{
Extract Year/Month/Day from date string
}
\examples{
D = c("2021-1-1", "2015/9/25")
year(D)
month(D)
day(D)
}
