% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{as_vec}
\alias{as_vec}
\title{Handy vector construction from string}
\usage{
as_vec(x)
}
\arguments{
\item{x}{String. A string to construct a vector from (see examples).
If \code{x} has length > 1, \code{stom::as_c_num} or \code{stom::as_c_chr}
would be called (depending on input type) to print out the
concatenation syntax \code{c(...)} in the console.}
}
\description{
Handy vector construction from string
}
\examples{
# Construct vector from string
as_vec("a b c")
as_vec("1 2 -3.1")  # Auto type casting to numeric vector
as_vec("4, 5, 6")
as_vec("4, 5  6")   # Commas have precedence over spaces

# Construct concatenation syntax from vector
as_vec(letters[1:3])
as_vec(1:3)
}
