% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{export_docs}
\alias{export_docs}
\alias{export_docs_pdf}
\alias{export_docs_html}
\title{Extract docstring from R script to various output formats}
\usage{
export_docs(path, outfp = NULL)

export_docs_pdf(path, outfp, style = "amsart")

export_docs_html(path, outfp)
}
\arguments{
\item{path}{Path to input R script.}

\item{outfp}{Output file path to write to.}

\item{style}{String. Predefined style to apply to PDF outputs. Currently,
only \code{amsart} is available.}
}
\description{
The docstring is defined as lines that start with \verb{#'}. To avoid problems
when converting to \code{.pdf} and \code{.html}, the format of documentation should
comply to Pandoc's markdown syntax.
To allow for the execution of  \code{export_docs_html()} and \code{export_docs_pdf()},
pandoc needs to be available through the command line. In addition, a LaTeX
distribution (e.g. TeXLive) is also required for PDF compilation.
}
\examples{
(fin = system.file("cases", "wine_network", "wine.R", package = "stom"))
docstring = export_docs(fin)  # Return as character vector if `outfp=NULL`
cat(docstring, sep="\n")
\dontrun{
export_docs(fin, "docs.md")
export_docs_pdf(fin, "docs.pdf")
export_docs_html(fin, "docs.html")
}
}
