% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract2.R
\name{extract2}
\alias{extract2}
\title{R6 representation of posterior samples}
\usage{
extract2(fit, ...)
}
\arguments{
\item{fit}{A \code{CmdStanFit} object passed to \code{\link[=extract]{extract()}}.}

\item{...}{Arguments passed to \code{\link[=extract]{extract()}}.}
}
\value{
An \code{R6}, \code{postParams} object.
}
\description{
\code{extract2()} provides an object-oriented interface to the posterior
samples extracted from a Stan model. In particular, high dimensional
parameters (i.e., vectors, matrices, and arrays) are gathered into arrays of
corresponding dimensions in R. These restructured high dimensional
parameters, as well as other scalar parameters, are exposed as public methods
in the returned R6 object. See below for an example.
}
\examples{
fp_mod = system.file("cases", "treatment_dynamics", "1-simplified-normal-model", "m1.RDS", package="stom")
m = readRDS(fp_mod)  # CmdStanFit model cache
post = extract2(m)
str(post)

post$B_AD() |> str()
post$B_TE() |> str()
post$E() |> str()
}
