% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{inline}
\alias{inline}
\alias{inline_zh}
\title{Inline Formatting of Vectors}
\usage{
inline(x, sep = ", ", last_sep = ", and ", two_sep = " and ")

inline_zh(x, sep = "、", last_sep = "、", two_sep = " 與 ")
}
\description{
Inline Formatting of Vectors
}
\examples{
inline(1:2)
inline(1:3)
inline_zh(1:2)
inline_zh(1:3)
}
