% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_format_conversion.R
\name{pdf2png}
\alias{pdf2png}
\alias{pdf2svg}
\alias{svg2pdf}
\alias{svg2png}
\title{Vector graphics conversion}
\usage{
pdf2png(
  infile,
  outfile = NULL,
  page = NULL,
  dpi = 400,
  white = FALSE,
  backend = c("mutool", "inkscape")
)

pdf2svg(
  infile,
  outfile = NULL,
  page = NULL,
  white = FALSE,
  backend = c("mutool", "inkscape")
)

svg2pdf(infile, outfile = NULL)

svg2png(infile, outfile = NULL, dpi = 1200)
}
\arguments{
\item{infile}{String. Path to input PDF file.}

\item{outfile}{String. By default, \code{outfile} is set to \code{NULL}, in which the
input PDF file stem is used.}

\item{page}{Integer vector. The pages (starting with 1) to extract from the
PDF file, defaults to \code{NULL}, in which every page is converted to
an image.}

\item{dpi}{Integer. "Dots Per Inch" for PNG outputs with \code{pdf2png()}.}

\item{backend}{String. The tool for converting images from PDFs. Defaults to \code{"mutool"}.
Currently supported options: \code{"mutool"} and \code{"inkscape"}}
}
\description{
Convert a PDF/SVG file to images of other formats with Inkscape.
}
\details{
To use this function, \code{mutool} (v1.24.0) or \code{inkscape} (v1.2.2) has to be
available from the command line. Alternatively, users can specify the
absolute path to inkscape by the option \code{stom.inkscape}
(e.g., \code{options(stom.inkscape = "/home/stom/bin/inkscape")}).
}
\examples{
\dontrun{
infile = system.file("inkscape", "dag.pdf", package = "stom")
pdf2svg( infile, outfile="dag.svg" )
pdf2png( infile, outfile="dag.png", dpi=800 )
svg2pdf( "dag.svg", "dag2.pdf" )
svg2png( "dag.svg", "dag2.png" )
}
}
