% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{precis}
\alias{precis}
\alias{extract}
\alias{erect}
\title{Posterior summary and extraction from CmdStanFit objects}
\usage{
precis(fit, depth = 1, pars = NULL, lp = F)

extract(fit, pars = NULL, lp = F)

erect(d, pars, col = "mean")
}
\arguments{
\item{fit}{A \code{CmdStanFit} object.}

\item{depth}{Integer. The level of parameters to display. \code{0} for scalar;
\code{1} for scalar and vector; \code{2} for scalar, vector, and matrix; and
so on.}

\item{pars}{Character vector. Names of the variables to include.
Passed to the \code{variables} argument in cmdstanr's
\verb{$summary()}/\verb{$draws()}.}

\item{lp}{Boolean. Whether to include log probabilities in output.}

\item{d}{data frame. Posterior summary data frame returned from \code{\link[=precis]{precis()}}.}
}
\value{
A data frame
}
\description{
Posterior summary and extraction from CmdStanFit objects
}
\examples{
# Fitted model
fp = system.file("cases", "wine_network", "wine2_normal_first_level2.RDS", package="stom")
m = readRDS(fp)

# Posterior distribution summary
d = precis( m, depth=2, pars="Int,J" )  # or pars=c("Int", "J")
d

# Reconstruct high-dimensional data structure from posterior means
erect( d, "Int" )
erect( d, "Int", "sd" )
erect( d, "Int,J" )

# Extract posterior samples
post = extract(m)
str(post)
}
