% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.R
\name{xclip}
\alias{xclip}
\alias{fclip}
\title{Copy to and from clipboard}
\usage{
xclip(x)

fclip()
}
\arguments{
\item{x}{Character.}
}
\description{
Copy to and from clipboard
}
\examples{
fp = 'C:\\\\Users\\\\rd\\\\bin\\\\export_docs_pdf.R'
fp |> xclip()  # copy to clipboard
fclip() |> cat()
}
